requirements.txt
requests
Image
python-docx
pytesseract